Code quality pass: clean /W3 C++ + ruff-clean Python
Runs the C++ plugin through MSVC /W3 and the Python runtime through
ruff, fixes every real issue reported.

C++ (/W3 clean):
- Bridge.cpp
    * SendAll/RecvAll now loop in <=1 MiB chunks cast to int, fixing
      C4267 size_t->int truncation warnings on Windows (recv/send take
      int, not size_t).
    * CMakeLists adds _CRT_SECURE_NO_WARNINGS and
      _WINSOCK_DEPRECATED_NO_WARNINGS so fopen/fscanf/getenv and
      inet_addr stop emitting C4996 nags. We use them deliberately.
- Render.cpp
    * Explicit static_cast<float>() on all PF_FpLong -> float slider
      reads (C4244). Previously implicit narrowing.

Python (ruff clean):
- Autofixed 54 issues across pyproject.toml scope: Optional[X] -> X|None,
  sorted imports, quoted annotations removed, OSError/TimeoutError
  aliases modernised.
- Removed channel-swap semicolon soup in handler.py
  (rgba[:,:,0] = ...; rgba[:,:,1] = ...) — one statement per line with
  inline channel labels.
- mlx/availability probe in hardware.py now imports mlx.core without
  the unused `as mx` alias (+ explicit `# noqa: F401` comment so the
  intent is obvious).
- `import torch.nn as nn` / `import torch.nn.functional as F` marked
  `# noqa: N812` — standard PyTorch convention, not an actual style
  violation.
- Long lines in main.py / hardware.py / handler.py / engines split
  into readable blocks so everything fits within the 100-char limit
  ruff.toml enforces.

Also strips the temporary per-step timing instrumentation that was
added in commit 768d3c9 to diagnose the cache-hash regression.

+++ plugin/src/CorridorKeyAE_Render.cpp
@@ -317,12 +317,14 @@ A_Err RenderEffect(
         request.pixel_data = ConvertTo8bpc(input, bpp);
         request.rowbytes = input->width * 4; // 8bpc, no padding
 
-        // Read effect parameters from AE
+        // Read effect parameters from AE. Slider values are PF_FpLong (double) —
+        // cast explicitly to float to match our FrameRequest fields and silence
+        // MSVC C4244 truncation warnings.
         request.output_mode = params[PARAM_OUTPUT_MODE]->u.pd.value - 1;
-        request.despill = params[PARAM_DESPILL_STRENGTH]->u.fs_d.value;
-        request.despeckle = params[PARAM_DESPECKLE_STRENGTH]->u.fs_d.value;
-        request.refiner = params[PARAM_REFINER_STRENGTH]->u.fs_d.value;
-        request.matte_cleanup = params[PARAM_MATTE_CLEANUP]->u.fs_d.value;
+        request.despill       = static_cast<float>(params[PARAM_DESPILL_STRENGTH]->u.fs_d.value);
+        request.despeckle     = static_cast<float>(params[PARAM_DESPECKLE_STRENGTH]->u.fs_d.value);
+        request.refiner       = static_cast<float>(params[PARAM_REFINER_STRENGTH]->u.fs_d.value);
+        request.matte_cleanup = static_cast<float>(params[PARAM_MATTE_CLEANUP]->u.fs_d.value);
 
         // Check out the alpha hint layer (if user selected one)
         PF_ParamDef hint_param;
@@ -492,12 +494,12 @@ A_Err SmartRender(
                               in_data->time_step, in_data->time_scale, &cleanup_param));
 
         if (!err) {
-            request.output_mode = mode_param.u.pd.value - 1;
-            request.quality_mode = quality_param.u.pd.value - 1;
-            request.despill = despill_param.u.fs_d.value;
-            request.despeckle = despeckle_param.u.fs_d.value;
-            request.refiner = refiner_param.u.fs_d.value;
-            request.matte_cleanup = cleanup_param.u.fs_d.value;
+            request.output_mode   = mode_param.u.pd.value - 1;
+            request.quality_mode  = quality_param.u.pd.value - 1;
+            request.despill       = static_cast<float>(despill_param.u.fs_d.value);
+            request.despeckle     = static_cast<float>(despeckle_param.u.fs_d.value);
+            request.refiner       = static_cast<float>(refiner_param.u.fs_d.value);
+            request.matte_cleanup = static_cast<float>(cleanup_param.u.fs_d.value);
         }
 
         ERR2(PF_CHECKIN_PARAM(in_data, &mode_param));
